requirements.txt
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
decorator==4.4.2
docx2pdf==0.1.8
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
moviepy==1.0.3
numpy==1.26.4
pillow==10.2.0
proglog==0.1.10
pyaes==1.6.1
Pyrogram==2.0.106
PySocks==1.7.1
python-telegram-bot==21.0.1
pytube==15.0.0
requests==2.31.0
sniffio==1.3.1
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
